% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/asku_knit.R
\name{asku}
\alias{asku}
\title{Hide your content in html}
\usage{
asku(content, Q = "What's the answer?", A, is.md = TRUE)
}
\arguments{
\item{content}{Character representing the content that will be hidden in
the generated html document.}

\item{Q}{Question promoted to the user.}

\item{A}{The correct answer expected.}

\item{is.md}{Is the content in markdown format or in raw html?}
}
\value{
A "shiny.tag" S3 object.
}
\description{
Hide your content in html
}
\examples{
frag <- asku(content = "## Header\\n\\n- list **item1**\\n- list *item2*", A = "42")
frag
htmltools::browsable(frag)
}
